requirements.txt
# =============================================================================
# OSINT News Aggregator - Python Dependencies
# =============================================================================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# OpenAI
openai>=2.0.0

# LangChain for agents and tools
langchain>=1.0.0
langchain-community>=0.4.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
langchain-tavily>=0.2.0
langgraph>=1.0.0
langgraph-checkpoint>=2.0.0  # Para checkpoints y human-in-the-loop

# LangSmith para observabilidad (opcional pero recomendado)
langsmith>=0.2.0

# Tavily Search API
tavily-python>=0.7.0

# HTTP requests
requests>=2.31.0
httpx>=0.28.0
anyio>=4.5.0

# DuckDuckGo Search (no API key required)
duckduckgo-search>=8.0.0

# HTML parsing for web scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# =============================================================================
# OSINT Tools - Modern replacements for OSRFramework
# =============================================================================

# Maigret - Username OSINT (replaces OSRFramework usufy)
# Searches for usernames across 500+ platforms
# https://github.com/soxoj/maigret
maigret>=0.4.4

# bbot - Attack surface enumeration
# Comprehensive subdomain/web reconnaissance tool
# https://github.com/blacklanternsecurity/bbot
bbot>=2.0.0

# =============================================================================

# Telegram integration
python-telegram-bot>=20.0

# Utilities
python-dateutil>=2.8.0
validators>=0.22.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
